liquidity_risk_monitor/requirements.txt
numpy==2.3.1
packaging==25.0
pandas==2.3.1
patsy==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
scipy==1.16.0
six==1.17.0
statsmodels==0.14.5
tzdata==2025.2
pandas
numpy
matplotlib
statsmodels
python-dotenv
